Moved BackToAppointments component.
