+++ src/applications/check-in/components/BackToAppointments.jsx
@@ -4,11 +4,11 @@ import PropTypes from 'prop-types';
 
 import recordEvent from 'platform/monitoring/record-event';
 
-import { createAnalyticsSlug } from '../../utils/analytics';
-import { useFormRouting } from '../../hooks/useFormRouting';
+import { createAnalyticsSlug } from '../utils/analytics';
+import { useFormRouting } from '../hooks/useFormRouting';
 import { withRouter } from 'react-router';
 
-import { URLS } from '../../utils/navigation';
+import { URLS } from '../utils/navigation';
 
 const BackToAppointments = ({ router, triggerRefresh }) => {
   const { jumpToPage } = useFormRouting(router);
