+++ src/applications/check-in/day-of/pages/Confirmation/MultipleAppointments.jsx
@@ -6,7 +6,7 @@ import { VaAlert } from 'web-components/react-bindings';
 import { focusElement } from 'platform/utilities/ui';
 
 import BackToHome from '../../components/BackToHome';
-import BackToAppointments from '../../components/BackToAppointments';
+import BackToAppointments from '../../../components/BackToAppointments';
 import Footer from '../../components/Footer';
 import AppointmentLocation from '../../components/AppointmentDisplay/AppointmentLocation';
 import TravelPayReimbursementLink from '../../../components/TravelPayReimbursementLink';
